+++ base/FighterManagerDlg.h
@@ -55,7 +55,7 @@ private slots:
 	//void on_buttonBox_accepted();
 	void on_pushButton_remove_pressed();
 	void on_pushButton_add_pressed();
-    void on_toolButton_openFolder_pressed();
+    void on_toolButton_openFile_pressed();
     void on_toolButton_reload_pressed();
     void on_toolButton_save_pressed();